+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Next
+
+- **[Fix]** Fix detection of internal URLs.
+
 # 0.1.0 (2018-10-18)
 
 - **[Feature]** First release